backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
python-dotenv>=1.0
django-cors-headers>=4.3
Pillow>=10.0
numpy>=1.26
opencv-python-headless>=4.10
google-genai>=0.6
google-generativeai>=0.3
celery>=5.3
redis>=5.0
rembg>=2.0.50
onnxruntime>=1.16.0
